build_requirements.txt
alabaster==0.7.9
babel==2.3.4
decorator==4.0.10
docutils==0.12
freezegun==0.3.8
imagesize==0.7.1
jinja2==2.8
markupsafe==0.23
networkx==1.11
numpydoc==0.6.0
py==1.4.31
pybuilder==0.11.9
pygments==2.1.3
pytest-asyncio==0.5.0
pytest-benchmark==3.0.0
pytest-capturelog==0.7
pytest-env==0.6.0
pytest-mock==1.5.0
pytest-timeout==1.2.0
pytest==3.0.4
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
snowballstemmer==1.2.1
sphinx-rtd-theme==0.1.9
sphinx==1.4.9
tblib==1.3.0
wheel==0.29.0

# The following packages are considered to be unsafe in a requirements file:
# pip
